requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# YOLOv8/v11 detection
ultralytics>=8.0.0

# Data processing
opencv-python>=4.8.0
numpy>=1.24.0
albumentations>=1.3.0
Pillow>=9.0.0

# API server
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Visualization and metrics
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
Pillow>=9.0.0

# Logging and monitoring
wandb>=0.15.0  # Optional: for experiment tracking

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.3.0
